+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Version X.Y.Z (YYYYMMDDNN)
 
 - Display attempt user email address in attempt report header section
+- Add `${email}` variable to filename patterns. Dots inside email addresses are automatically replaced with underscores.
 
 
 ## Version 5.0.0 (2026081100)